[buildkite] Unblock Jobs from whitelisted authors
We used to allow CI to be triggered if a PR was created by a whitelisted
author (manually whitelisted or member of the GitHub repo), but recently
moved to using Buildkite's blocked steps instead so that we could make
the pipelines public. When doing so we locked users without a Buildkite
account but who have write access to the repo or are otherwise trusted
from triggering CI.

This PR unblocks a Buildkite job once it has updated the GitHub status
for a commit. It is currently only based on whitelisting, since
labelling a PR from a new external contributor is relatively as much
effort as just opening the associated Buildkite build and unblocking it,
and makes the code simpler.